+++ lib/Db/SettingTemplateMapper.php
@@ -30,7 +30,23 @@ class SettingTemplateMapper extends QBMapper {
 
 		return $this->findEntity($qb);
 	}
+	/**
+	 * @throws \OCP\AppFramework\Db\DoesNotExistException if not found
+	 * @throws \OCP\AppFramework\Db\MultipleObjectsReturnedException if more than one result
+	 *
+	 * @return \OCP\AppFramework\Db\Entity
+	 */
+	public function getByTemplateKey(string $templatekey): \OCP\AppFramework\Db\Entity {
+		$qb = $this->db->getQueryBuilder();
 
+		$qb->select('*')
+		   ->from('sndnt_stngtmplt')
+		   ->where(
+			   $qb->expr()->eq('templatekey', $qb->createNamedParameter($templatekey, IQueryBuilder::PARAM_STR))
+		   );
+
+		return $this->findEntity($qb);
+	}
 	/**
 	 * @return \OCP\AppFramework\Db\Entity[]
 	 *
@@ -47,13 +63,13 @@ class SettingTemplateMapper extends QBMapper {
 		return $this->findEntities($qb);
 	}
 
-	public function settingKeyCount($name) {
+	public function settingTemplateCount($templatekey) {
 		$qb = $this->db->getQueryBuilder();
 
 		$qb->selectAlias($qb->createFunction('COUNT(*)'), 'count')
 		   ->from('sndnt_stngtmplt')
 		   ->where(
-			   $qb->expr()->eq('templatename', $qb->createNamedParameter($name, IQueryBuilder::PARAM_STR))
+			   $qb->expr()->eq('templatekey', $qb->createNamedParameter($templatekey, IQueryBuilder::PARAM_STR))
 		   );
 
 		$cursor = $qb->execute();